make api independent from each other
this will make all api could be called using `async with` style rather
than dependency with the BMKG class. this also make api more flexible.
