+++ src/bmkg/api/api.py
@@ -1,3 +1,7 @@
+from traceback import TracebackException
+from types import TracebackType
+from typing import Self
+
 from aiohttp import ClientSession
 
 __all__ = ["API"]
@@ -12,3 +16,17 @@ class API:
         self._session = (
             session if session is not None else ClientSession("https://data.bmkg.go.id")
         )
+
+    async def __aenter__(self) -> Self:
+        return self
+
+    async def __aexit__(
+        self,
+        exc_type: Exception,
+        exc_val: TracebackException,
+        traceback: TracebackType,
+    ) -> None:
+        await self.close()
+
+    async def close(self) -> None:
+        await self._session.close()
