+++ src/bmkg/api/bmkg.py
@@ -1,6 +1,4 @@
-from traceback import TracebackException
-from types import TracebackType
-from typing import Self
+from aiohttp import ClientSession
 
 from .api import API
 from .earthquake import Earthquake
@@ -18,22 +16,8 @@ class BMKG(API):
         weather_forecast (WeatherForecast): weather forecast api interface.
     """
 
-    def __init__(self) -> None:
-        API.__init__(self)
+    def __init__(self, session: ClientSession | None = None) -> None:
+        API.__init__(self, session)
 
         self.earthquake = Earthquake(self._session)
         self.weather_forecast = WeatherForecast(self._session)
-
-    async def __aenter__(self) -> Self:
-        return self
-
-    async def __aexit__(
-        self,
-        exc_type: Exception,
-        exc_val: TracebackException,
-        traceback: TracebackType,
-    ) -> None:
-        await self.close()
-
-    async def close(self) -> None:
-        await self._session.close()
